requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Janome==0.3.2
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.0
PyYAML==3.12
Werkzeug==0.12.2
